+++ source/Halibut.Tests/PollingClientConnectionHandlingFixture.cs
@@ -6,6 +6,7 @@ using FluentAssertions;
 using Halibut.ServiceModel;
 using Halibut.Tests.TestServices;
 using Halibut.Tests.Util;
+using Halibut.Tests.Util.TcpUtils;
 using NUnit.Framework;
 
 namespace Halibut.Tests
