+++ source/Halibut.Tests/Util/TcpUtils/TCPListenerWhichKillsNewConnections.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Serilog;
 
-namespace Halibut.Tests.Util
+namespace Halibut.Tests.Util.TcpUtils
 {
     
     /// <summary>
