+++ source/Halibut.Tests/WhenTheTcpConnectionStopsSendingData.cs
@@ -5,6 +5,7 @@ using Halibut.Diagnostics;
 using Halibut.ServiceModel;
 using Halibut.Tests.TestServices;
 using Halibut.Tests.Util;
+using Halibut.Tests.Util.TcpUtils;
 using NUnit.Framework;
 
 namespace Halibut.Tests
